@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: aserto/authorizer/v2/api/decision_logs.proto
 
@@ -395,7 +395,7 @@ func file_aserto_authorizer_v2_api_decision_logs_proto_rawDescGZIP() []byte {
 }
 
 var file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_aserto_authorizer_v2_api_decision_logs_proto_goTypes = []interface{}{
+var file_aserto_authorizer_v2_api_decision_logs_proto_goTypes = []any{
 	(*Decision)(nil),              // 0: aserto.authorizer.v2.api.Decision
 	(*DecisionUser)(nil),          // 1: aserto.authorizer.v2.api.DecisionUser
 	(*DecisionPolicy)(nil),        // 2: aserto.authorizer.v2.api.DecisionPolicy
@@ -433,7 +433,7 @@ func file_aserto_authorizer_v2_api_decision_logs_proto_init() {
 	file_aserto_authorizer_v2_api_identity_context_proto_init()
 	file_aserto_authorizer_v2_api_policy_instance_proto_init()
 	if !protoimpl.UnsafeEnabled {
-		file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Decision); i {
 			case 0:
 				return &v.state
@@ -445,7 +445,7 @@ func file_aserto_authorizer_v2_api_decision_logs_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*DecisionUser); i {
 			case 0:
 				return &v.state
@@ -457,7 +457,7 @@ func file_aserto_authorizer_v2_api_decision_logs_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*DecisionPolicy); i {
 			case 0:
 				return &v.state
@@ -470,7 +470,7 @@ func file_aserto_authorizer_v2_api_decision_logs_proto_init() {
 			}
 		}
 	}
-	file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[0].OneofWrappers = []interface{}{}
+	file_aserto_authorizer_v2_api_decision_logs_proto_msgTypes[0].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
