@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: aserto/authorizer/v2/api/identity_context.proto
 
@@ -182,7 +182,7 @@ func file_aserto_authorizer_v2_api_identity_context_proto_rawDescGZIP() []byte {
 
 var file_aserto_authorizer_v2_api_identity_context_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_aserto_authorizer_v2_api_identity_context_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_aserto_authorizer_v2_api_identity_context_proto_goTypes = []interface{}{
+var file_aserto_authorizer_v2_api_identity_context_proto_goTypes = []any{
 	(IdentityType)(0),       // 0: aserto.authorizer.v2.api.IdentityType
 	(*IdentityContext)(nil), // 1: aserto.authorizer.v2.api.IdentityContext
 }
@@ -201,7 +201,7 @@ func file_aserto_authorizer_v2_api_identity_context_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_aserto_authorizer_v2_api_identity_context_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_api_identity_context_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*IdentityContext); i {
 			case 0:
 				return &v.state
