@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: aserto/authorizer/v2/api/module.proto
 
@@ -146,7 +146,7 @@ func file_aserto_authorizer_v2_api_module_proto_rawDescGZIP() []byte {
 }
 
 var file_aserto_authorizer_v2_api_module_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_aserto_authorizer_v2_api_module_proto_goTypes = []interface{}{
+var file_aserto_authorizer_v2_api_module_proto_goTypes = []any{
 	(*Module)(nil),         // 0: aserto.authorizer.v2.api.Module
 	(*structpb.Value)(nil), // 1: google.protobuf.Value
 }
@@ -165,7 +165,7 @@ func file_aserto_authorizer_v2_api_module_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_aserto_authorizer_v2_api_module_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_api_module_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Module); i {
 			case 0:
 				return &v.state
@@ -178,7 +178,7 @@ func file_aserto_authorizer_v2_api_module_proto_init() {
 			}
 		}
 	}
-	file_aserto_authorizer_v2_api_module_proto_msgTypes[0].OneofWrappers = []interface{}{}
+	file_aserto_authorizer_v2_api_module_proto_msgTypes[0].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
