@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: aserto/authorizer/v2/api/policy_context.proto
 
@@ -109,7 +109,7 @@ func file_aserto_authorizer_v2_api_policy_context_proto_rawDescGZIP() []byte {
 }
 
 var file_aserto_authorizer_v2_api_policy_context_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_aserto_authorizer_v2_api_policy_context_proto_goTypes = []interface{}{
+var file_aserto_authorizer_v2_api_policy_context_proto_goTypes = []any{
 	(*PolicyContext)(nil), // 0: aserto.authorizer.v2.api.PolicyContext
 }
 var file_aserto_authorizer_v2_api_policy_context_proto_depIdxs = []int32{
@@ -126,7 +126,7 @@ func file_aserto_authorizer_v2_api_policy_context_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_aserto_authorizer_v2_api_policy_context_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_api_policy_context_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*PolicyContext); i {
 			case 0:
 				return &v.state
