@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: aserto/authorizer/v2/authorizer.proto
 
@@ -1641,7 +1641,7 @@ func file_aserto_authorizer_v2_authorizer_proto_rawDescGZIP() []byte {
 
 var file_aserto_authorizer_v2_authorizer_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
 var file_aserto_authorizer_v2_authorizer_proto_msgTypes = make([]protoimpl.MessageInfo, 17)
-var file_aserto_authorizer_v2_authorizer_proto_goTypes = []interface{}{
+var file_aserto_authorizer_v2_authorizer_proto_goTypes = []any{
 	(PathSeparator)(0),            // 0: aserto.authorizer.v2.PathSeparator
 	(TraceLevel)(0),               // 1: aserto.authorizer.v2.TraceLevel
 	(*InfoRequest)(nil),           // 2: aserto.authorizer.v2.InfoRequest
@@ -1731,7 +1731,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*InfoRequest); i {
 			case 0:
 				return &v.state
@@ -1743,7 +1743,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*InfoResponse); i {
 			case 0:
 				return &v.state
@@ -1755,7 +1755,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetPolicyRequest); i {
 			case 0:
 				return &v.state
@@ -1767,7 +1767,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetPolicyResponse); i {
 			case 0:
 				return &v.state
@@ -1779,7 +1779,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ListPoliciesRequest); i {
 			case 0:
 				return &v.state
@@ -1791,7 +1791,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*ListPoliciesResponse); i {
 			case 0:
 				return &v.state
@@ -1803,7 +1803,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*DecisionTreeRequest); i {
 			case 0:
 				return &v.state
@@ -1815,7 +1815,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*DecisionTreeOptions); i {
 			case 0:
 				return &v.state
@@ -1827,7 +1827,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*DecisionTreeResponse); i {
 			case 0:
 				return &v.state
@@ -1839,7 +1839,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*IsRequest); i {
 			case 0:
 				return &v.state
@@ -1851,7 +1851,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*Decision); i {
 			case 0:
 				return &v.state
@@ -1863,7 +1863,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*IsResponse); i {
 			case 0:
 				return &v.state
@@ -1875,7 +1875,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*QueryOptions); i {
 			case 0:
 				return &v.state
@@ -1887,7 +1887,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*QueryRequest); i {
 			case 0:
 				return &v.state
@@ -1899,7 +1899,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*CompileRequest); i {
 			case 0:
 				return &v.state
@@ -1911,7 +1911,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*CompileResponse); i {
 			case 0:
 				return &v.state
@@ -1923,7 +1923,7 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 				return nil
 			}
 		}
-		file_aserto_authorizer_v2_authorizer_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
+		file_aserto_authorizer_v2_authorizer_proto_msgTypes[16].Exporter = func(v any, i int) any {
 			switch v := v.(*QueryResponse); i {
 			case 0:
 				return &v.state
@@ -1936,12 +1936,12 @@ func file_aserto_authorizer_v2_authorizer_proto_init() {
 			}
 		}
 	}
-	file_aserto_authorizer_v2_authorizer_proto_msgTypes[2].OneofWrappers = []interface{}{}
-	file_aserto_authorizer_v2_authorizer_proto_msgTypes[4].OneofWrappers = []interface{}{}
-	file_aserto_authorizer_v2_authorizer_proto_msgTypes[6].OneofWrappers = []interface{}{}
-	file_aserto_authorizer_v2_authorizer_proto_msgTypes[9].OneofWrappers = []interface{}{}
-	file_aserto_authorizer_v2_authorizer_proto_msgTypes[13].OneofWrappers = []interface{}{}
-	file_aserto_authorizer_v2_authorizer_proto_msgTypes[14].OneofWrappers = []interface{}{}
+	file_aserto_authorizer_v2_authorizer_proto_msgTypes[2].OneofWrappers = []any{}
+	file_aserto_authorizer_v2_authorizer_proto_msgTypes[4].OneofWrappers = []any{}
+	file_aserto_authorizer_v2_authorizer_proto_msgTypes[6].OneofWrappers = []any{}
+	file_aserto_authorizer_v2_authorizer_proto_msgTypes[9].OneofWrappers = []any{}
+	file_aserto_authorizer_v2_authorizer_proto_msgTypes[13].OneofWrappers = []any{}
+	file_aserto_authorizer_v2_authorizer_proto_msgTypes[14].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
