@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.4.0
 // - protoc             (unknown)
 // source: aserto/authorizer/v2/authorizer.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.62.0 or later.
+const _ = grpc.SupportPackageIsVersion8
 
 const (
 	Authorizer_DecisionTree_FullMethodName = "/aserto.authorizer.v2.Authorizer/DecisionTree"
@@ -50,8 +50,9 @@ func NewAuthorizerClient(cc grpc.ClientConnInterface) AuthorizerClient {
 }
 
 func (c *authorizerClient) DecisionTree(ctx context.Context, in *DecisionTreeRequest, opts ...grpc.CallOption) (*DecisionTreeResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DecisionTreeResponse)
-	err := c.cc.Invoke(ctx, Authorizer_DecisionTree_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Authorizer_DecisionTree_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,8 +60,9 @@ func (c *authorizerClient) DecisionTree(ctx context.Context, in *DecisionTreeReq
 }
 
 func (c *authorizerClient) Is(ctx context.Context, in *IsRequest, opts ...grpc.CallOption) (*IsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(IsResponse)
-	err := c.cc.Invoke(ctx, Authorizer_Is_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Authorizer_Is_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -68,8 +70,9 @@ func (c *authorizerClient) Is(ctx context.Context, in *IsRequest, opts ...grpc.C
 }
 
 func (c *authorizerClient) Query(ctx context.Context, in *QueryRequest, opts ...grpc.CallOption) (*QueryResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(QueryResponse)
-	err := c.cc.Invoke(ctx, Authorizer_Query_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Authorizer_Query_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -77,8 +80,9 @@ func (c *authorizerClient) Query(ctx context.Context, in *QueryRequest, opts ...
 }
 
 func (c *authorizerClient) Compile(ctx context.Context, in *CompileRequest, opts ...grpc.CallOption) (*CompileResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CompileResponse)
-	err := c.cc.Invoke(ctx, Authorizer_Compile_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Authorizer_Compile_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -86,8 +90,9 @@ func (c *authorizerClient) Compile(ctx context.Context, in *CompileRequest, opts
 }
 
 func (c *authorizerClient) ListPolicies(ctx context.Context, in *ListPoliciesRequest, opts ...grpc.CallOption) (*ListPoliciesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListPoliciesResponse)
-	err := c.cc.Invoke(ctx, Authorizer_ListPolicies_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Authorizer_ListPolicies_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -95,8 +100,9 @@ func (c *authorizerClient) ListPolicies(ctx context.Context, in *ListPoliciesReq
 }
 
 func (c *authorizerClient) GetPolicy(ctx context.Context, in *GetPolicyRequest, opts ...grpc.CallOption) (*GetPolicyResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GetPolicyResponse)
-	err := c.cc.Invoke(ctx, Authorizer_GetPolicy_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Authorizer_GetPolicy_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -104,8 +110,9 @@ func (c *authorizerClient) GetPolicy(ctx context.Context, in *GetPolicyRequest, 
 }
 
 func (c *authorizerClient) Info(ctx context.Context, in *InfoRequest, opts ...grpc.CallOption) (*InfoResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(InfoResponse)
-	err := c.cc.Invoke(ctx, Authorizer_Info_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Authorizer_Info_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
